empleados and compra tabs now shows registers

+++ src/main/java/Logica/Clases/Empleado.java
@@ -6,7 +6,7 @@ public class Empleado extends Persona {
     private String fechaIngreso;
 
 
-    public Empleado(int dni, String nombre, String direccion, long telefono, int nroLegajo, String fechaIngreso){
+    public Empleado(long dni, String nombre, String direccion, long telefono, int nroLegajo, String fechaIngreso){
         super(dni, nombre, direccion, telefono);
         this.nroLegajo = nroLegajo;
         this.fechaIngreso = fechaIngreso;
